requirements.txt
alembic==1.4.2
arrow==0.15.8
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Dance==3.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
oauthlib==3.1.0
pendulum==2.1.2
psycopg2-binary==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytzdata==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rq==1.5.0
rq-dashboard==0.6.1
sentry-sdk==0.16.2
six==1.15.0
SQLAlchemy==1.3.18
SQLAlchemy-Utils==0.36.8
urllib3==1.25.10
URLObject==2.4.3
validators==0.16.0
Werkzeug==1.0.1
WTForms==2.3.1
